requirements.txt
numpy>=2.3.3
pandas>=2.3.2
tensorflow>=2.20.0
keras>=3.11.3
pyyaml>=6.0.3
python>=3.12.1